# lib/api-zod/src/generated/types/clubEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ClubEntry {
  id: number;
  name: string;
  logo: string;
  league: string;
  leagueId: number;
  country?: string;
}