# lib/api-zod/src/generated/types/putClubsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ClubEntry } from "./clubEntry";

export interface PutClubsBody {
  clubs: ClubEntry[];
  cachedAt: number;
}