# lib/api-zod/src/generated/types/putSquadBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SquadPlayer } from "./squadPlayer";

export interface PutSquadBody {
  players: SquadPlayer[];
  source: string;
  cachedAt: number;
}